requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
streamlit==1.44.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
python-dotenv==1.1.0
pandas==2.2.3
openai==1.76.0
transformers==4.51.3
torch==2.7.0
matplotlib==3.10.1
plotly==6.0.1
